pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "boardroom-scripts"
version = "0.0.0"
description = "Boardroom as plain scripts (not packaged)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7.0",
  "langgraph>=0.2.0",
  "notion-client>=2.3.0",
  "python-dotenv>=1.0.1",
  "pytest>=8.2.0",
  "openai>=1.12.0",
]

[tool.uv]
package = false
